+++ packages/react-magma-dom/src/components/Select/SelectTriggerButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { inputBaseStyles } from '../InputBase';
-import { CaretDownIcon } from 'react-magma-icons';
+import { defaultComponents, SelectComponents } from '../Select/components';
 import { ThemeContext } from '../../theme/ThemeContext';
 import styled from '../../theme/styled';
 
@@ -25,6 +25,7 @@ const ChildrenContainer = styled.div`
 interface SelectTriggerButtonInterface<T> {
   ariaDescribedBy?: string;
   children: React.ReactNode[];
+  customComponents?: SelectComponents;
   hasError?: boolean;
   disabled?: boolean;
   isInverse?: boolean;
@@ -37,6 +38,7 @@ export function SelectTriggerButton<T>(props: SelectTriggerButtonInterface<T>) {
   const {
     ariaDescribedBy,
     children,
+    customComponents,
     hasError,
     disabled,
     isInverse,
@@ -44,6 +46,10 @@ export function SelectTriggerButton<T>(props: SelectTriggerButtonInterface<T>) {
     toggleButtonProps,
   } = props;
   const theme = React.useContext(ThemeContext);
+
+  const { DropdownIndicator } = defaultComponents({
+    ...customComponents,
+  });
 
   return (
     <StyledButton
@@ -59,7 +65,7 @@ export function SelectTriggerButton<T>(props: SelectTriggerButtonInterface<T>) {
       tabIndex={disabled ? undefined : 0}
     >
       <ChildrenContainer>{children}</ChildrenContainer>
-      <CaretDownIcon size={10} style={{ flexShrink: 0 }} testId="caretDown" />
+      <DropdownIndicator />
     </StyledButton>
   );
 }