+++ packages/react-magma-dom/src/index.ts
@@ -1,7 +1,12 @@
 export { Alert, AlertProps } from './components/Alert';
 export { AlertVariant } from './components/AlertBase';
 export { Announce, AnnounceProps } from './components/Announce';
-export { Badge, BadgeColor, BadgeProps } from './components/Badge';
+export {
+  Badge,
+  BadgeColor,
+  BadgeProps,
+  BadgeVariant,
+} from './components/Badge';
 export { Banner, BannerProps } from './components/Banner';
 export { Breadcrumb, BreadcrumbProps } from './components/Breadcrumb';
 export { BreadcrumbItem } from './components/Breadcrumb/Item';
@@ -64,7 +69,6 @@ export {
 } from './components/InputBase';
 export { InputMessage } from './components/Input/InputMessage';
 export { Label } from './components/Label';
-export { LinkButton } from './components/LinkButton';
 export {
   LoadingIndicator,
   LoadingIndicatorProps,