+++ CHANGELOG.md
@@ -3,6 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.119](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.118...v2.0.0-beta.119) (2020-11-09)
+
+
+### Bug Fixes
+
+* **card:** fix heading styles ([bdd9598](http://stash.cengage.com:7999/FRONT/react-magma/commits/bdd95984b7c88f1be19f867574f8672228169de2))
+
+
+### Features
+
+* **select:** allow for custom dropdown indicator ([45f2d0d](http://stash.cengage.com:7999/FRONT/react-magma/commits/45f2d0d3e27ed53185bbc2511180657e57ffa8e4))
+
+
+
+
+
+# [2.0.0-beta.118](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.117...v2.0.0-beta.118) (2020-11-09)
+
+
+### Bug Fixes
+
+* **icons:** don't extract props for icon render ([0f71d98](http://stash.cengage.com:7999/FRONT/react-magma/commits/0f71d98db6979a96e48c1623809d3c42d127de44))
+
+
+
+
+
+# [2.0.0-beta.117](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.116...v2.0.0-beta.117) (2020-11-09)
+
+
+### Bug Fixes
+
+* **badge:** export badgeVariant enum ([ac8303b](http://stash.cengage.com:7999/FRONT/react-magma/commits/ac8303b2fe0c02901f711c62fcc80303c2822b4f))
+* **docs:** don't link local legacy selects ([0fed32f](http://stash.cengage.com:7999/FRONT/react-magma/commits/0fed32f21cc6cbcdd6e2c20e5f2aee081d99f0d2))
+* **docs:** fix linking of icons/legacy-selects ([e07b388](http://stash.cengage.com:7999/FRONT/react-magma/commits/e07b38849e9e89a260e087879b6bd6050f30563e))
+* **form-group:** make helperMessage prop optional ([0b3076a](http://stash.cengage.com:7999/FRONT/react-magma/commits/0b3076ad0e3673c8916e7b2ebfe6024c43e58c7d))
+* **icons:** always force bootstrap linking ([edc3135](http://stash.cengage.com:7999/FRONT/react-magma/commits/edc313562ed76c4a0cee92be5a85409bf0885e58))
+
+
+
+
+
+# [2.0.0-beta.116](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.115...v2.0.0-beta.116) (2020-11-03)
+
+**Note:** Version bump only for package react-magma
+
+
+
+
+
+# [2.0.0-beta.115](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.114...v2.0.0-beta.115) (2020-11-02)
+
+
+### Features
+
+* **card:** give inverse cards default bgcolor, add context for inverse ([4549420](http://stash.cengage.com:7999/FRONT/react-magma/commits/4549420d06df33d9905f8d35975ed3705b531bb0))
+
+
+
+
+
+# [2.0.0-beta.114](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.113...v2.0.0-beta.114) (2020-10-30)
+
+
+### Features
+
+* **linkbutton:** remove link button component ([99fe51f](http://stash.cengage.com:7999/FRONT/react-magma/commits/99fe51f7fe34f75a62d5b7d5ff89e02b933ac677))
+
+
+### BREAKING CHANGES
+
+* **linkbutton:** removed LinkButton component
+
+
+
+
+
 # [2.0.0-beta.113](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.112...v2.0.0-beta.113) (2020-10-29)