+++ packages/react-magma-icons/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.118](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.117...v2.0.0-beta.118) (2020-11-09)
+
+### Bug Fixes
+
+- **icons:** don't extract props for icon render ([0f71d98](http://stash.cengage.com:7999/FRONT/react-magma/commits/0f71d98db6979a96e48c1623809d3c42d127de44))
+
 # [2.0.0-beta.113](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.112...v2.0.0-beta.113) (2020-10-29)
 
 **Note:** Version bump only for package react-magma-icons